[REFACTOR]: 현재 시간 반환 함수 분리
